Update docker build
Switch to a Debian release where an OpenSSL squid package is available,
to use with SSLBump.

Add CA certifiates to the container so that the proxy itself can connect
to HTTPS endpoints.

Update the entrypoint script to add prepare an SSL DB that is also
needed for SSLBump and for tailing the default logs to output them to
stdout. Should be fixed later by setting a USER in the Dockerfile.