Update image to run as non root
USER directive added to dockerfile to run as squid default user.

To make this work a directory and a default configuration is needed for
the pid file.

Once the image is running as non root it is also possible to log to
stdout without issues, which is what we want with docker images so there
is no need to colloct any logs from within the container.
A defaul configration also added for the logging.

When a custom confiration is mounted it will need to include the
`/etc/squid/conf.d` or manually set the required configuration. See
`./conf.d/docker.conf`.

entrypoint.sh updated accordingly.